(BSR)[API] refactor: move ubble transactional emails to their own folder
